Added 3 Single DAL 3 Single Values on Dashboard including the % changes with rounded values Y-axis Label title was changed for TotalSalesBranch Y-axis rounded values - changed for TotalSalesBranch added a table for the EmployeeKPI page with Filtering (not functional) Changed all the layers from decimal to double Changed the css file to include css for kpiheader

+++ Test/Models/Interface/Financial/ITotalSalesSingleRepo.cs
@@ -8,6 +8,6 @@ namespace Test.Models.Interface.Financial
 {
     public interface ITotalSalesSingleRepo: IDisposable
     {
-        decimal? usp_TotalSalesSingle(int companyRef);
+        double? usp_TotalSalesSingle(int companyRef);
     }
 }
